+++ routes/user.ts
@@ -11,7 +11,7 @@
  */
 import express from 'express';
 
-import User from '../model/user';
+import User from '../models/user';
 
 const router = express.Router();
